third_party/blink/web_tests/platform/generic/http/tests/websocket/sub-protocol-expected.txt

Test WebSocket handshake success with sub protocol.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected
PASS protocol is "sub-protocol"
PASS successfullyParsed is true

TEST COMPLETE
